+++ src/main/kotlin/datamodel/hilfsclass/RoutineEvent.kt
@@ -3,7 +3,7 @@ package datamodel.hilfsclass
 import datamodel.function
 
 data class RoutineEvent(
-    var deviceId: Int,
-    var functionId: Int,
+    var deviceId: Long,
+    var functionId: Long,
     var funct: function,
 )
